Make MediatorConfiguration an interface
